+++ poisson_approval/__init__.py
@@ -108,6 +108,9 @@ from poisson_approval.random_factories.RandTauVectorUniform import RandTauVector
 
 # Meta-analysis
 from poisson_approval.meta_analysis.NiceStatsProfileOrdinal import NiceStatsProfileOrdinal
+from poisson_approval.meta_analysis.binary_plots import BinaryAxesSubplotPoisson, binary_figure
+from poisson_approval.meta_analysis.binary_shortcuts import binary_plot_n_equilibria, \
+    binary_plot_winners_at_equilibrium, binary_plot_winning_frequencies, XyyToProfile
 from poisson_approval.meta_analysis.ternary_plots import TernaryAxesSubplotPoisson, ternary_figure
 from poisson_approval.meta_analysis.ternary_shortcuts import ternary_plot_n_equilibria, \
     ternary_plot_winners_at_equilibrium, ternary_plot_winning_frequencies, SimplexToProfile