+++ poisson_approval/meta_analysis/ternary_plots.py
@@ -1,134 +1,10 @@
-import numpy as np
 import ternary
 from collections import Counter
 from ternary.helpers import simplex_iterator
 from matplotlib import pyplot as plt
 from matplotlib.patches import Patch
 from poisson_approval.meta_analysis.ternary_condorcet import draw_condorcet_zones
-
-
-_WHITE = np.array([1., 1., 1.])
-_BRIGHTNESS = 0.5
-
-
-def _abc_to_rgb(abc):
-    """Conversion from candidates to RGB color
-
-    Parameters
-    ----------
-    abc : tuple, list or ndarray
-        Array of size 3. Each coefficient is associated to a candidate a, b, c. Typically the coefficients sum to 1
-        or less than 1.
-
-    Returns
-    -------
-    ndarray
-        Array of size 3. RGB color. Roughly speaking, candidate `a` is associated to red, `b` to green and `c` to blue,
-        but we try to mix colors more reasonably than just an average in RGB space.
-
-    Examples
-    --------
-        >>> _abc_to_rgb([1, 0, 0])
-        array([1. , 0.5, 0.5])
-        >>> _abc_to_rgb([.5, .5, 0])
-        array([1. , 1. , 0.5])
-        >>> _abc_to_rgb([0., 0., 0])
-        array([0.5, 0.5, 0.5])
-    """
-    norm_inf = np.max(abc)
-    if norm_inf == 0:
-        rgb = np.zeros(3)
-    else:
-        norm_1 = np.sum(abc)
-        rgb = np.array(abc) * np.min([norm_1, 1]) / norm_inf
-    return rgb * (1 - _BRIGHTNESS) + _WHITE * _BRIGHTNESS
-
-
-_COLOR_NO_ONE = _abc_to_rgb([0, 0, 0])
-_COLOR_A = _abc_to_rgb([1, 0, 0])
-_COLOR_B = _abc_to_rgb([0, 1, 0])
-_COLOR_C = _abc_to_rgb([0, 0, 1])
-_COLOR_AB = _abc_to_rgb([1, 1, 0])
-_COLOR_AC = _abc_to_rgb([1, 0, 1])
-_COLOR_BC = _abc_to_rgb([0, 1, 1])
-_COLOR_ABC = _abc_to_rgb([1, 1, 1])
-
-
-def _uni_mix(color_1, color_2, weight_1, weight_2):
-    """Mix two colors, with a fade that emphasizes the two base colors.
-
-    Parameters
-    ----------
-    color_1, color_2 : tuple, list or ndarray
-        Arrays of size 3 in format RGB.
-    weight_1, weight_2 : Number
-        Respective weights of the two colors.
-
-    Returns
-    -------
-    ndarray
-        Array of size 3 in format RGB. The mix is not linear. Currently, the resulting color is a barycenter
-        with respective weights ``weight_1**n`` and ``weight_2**n``, with ``n = 2.5``. This makes a "sharper"
-        transition, so that we see more of the base colors and less of the mixes.
-
-    Examples
-    --------
-        >>> import numpy as np
-        >>> _uni_mix(color_1=[1, 0, 0], color_2=[0, 1, 0], weight_1=1, weight_2=0)
-        array([1., 0., 0.])
-        >>> _uni_mix(color_1=[1, 0, 0], color_2=[0, 1, 0], weight_1=.5, weight_2=.5)
-        array([0.5, 0.5, 0. ])
-        >>> _uni_mix(color_1=[1, 0, 0], color_2=[0, 1, 0], weight_1=0, weight_2=1)
-        array([0., 1., 0.])
-    """
-    n = 2.5
-    return (np.array(color_1) * weight_1 ** n + np.array(color_2) * weight_2 ** n) / (weight_1 ** n + weight_2 ** n)
-
-
-def _palette(x, y):
-    """Compute the data for a nice color palette.
-
-    Parameters
-    ----------
-    x, y : Number
-        Cartesian coordinates of the point.
-
-    Returns
-    -------
-    ndarray
-        Array of size 3 in format RGB. The resulting figure consists of:
-
-        * A nice color palette in the unit disk,
-        * A black circle on the unit circle,
-        * White background outside.
-
-    Examples
-    --------
-        >>> _palette(x=0, y=1.5)
-        array([1., 1., 1.])
-        >>> _palette(x=0, y=0.99)
-        array([0., 0., 0.])
-        >>> color_almost_a = _palette(x=0, y=0.95)
-        >>> np.all(np.isclose(color_almost_a, [1, .5, .5], atol=1E-2))
-        True
-    """
-    z = x + 1j * y
-    r = np.abs(z)
-    if r > 1:
-        return np.array([1., 1., 1.])
-    if r > 0.96:
-        return np.array([0., 0., 0.])
-    theta = (np.angle(z, 'deg') + 30) % 360 - 30
-    for theta_sup, color_inf, color_sup in zip([30, 90, 150, 210, 270, 330],
-                                               [_COLOR_C, _COLOR_AC, _COLOR_A, _COLOR_AB, _COLOR_B, _COLOR_BC],
-                                               [_COLOR_AC, _COLOR_A, _COLOR_AB, _COLOR_B, _COLOR_BC, _COLOR_C]):
-        if theta <= theta_sup:
-            weight_inf = (theta_sup - theta) / 60
-            color = _uni_mix(color_sup, color_inf, 1 - weight_inf, weight_inf)
-            return _uni_mix(color, _COLOR_ABC, r ** .8, 1 - r ** .8)
-
-
-_DATA_PALETTE = np.array([[_palette(x / 100, y / 100) for x in range(-100, 100)] for y in range(-100, 100)])
+from poisson_approval.meta_analysis.colors import *
 
 
 def _generate_heatmap_data(f, scale):
@@ -159,7 +35,7 @@ def _generate_heatmap_data(f, scale):
     d = dict()
     for (right, top, left) in simplex_iterator(scale):
         values = f(right / scale, top / scale, left / scale)
-        color = _abc_to_rgb(values)
+        color = abc_to_rgb(values)
         d[(right, top, left)] = (float(color[0]), float(color[1]), float(color[2]), 1.)
     return d
 
@@ -510,9 +386,9 @@ class TernaryAxesSubplotPoisson(ternary.TernaryAxesSubplot):  # pragma: no cover
         """
         def _edge_color(color):
             return [max(x - .4, 0) for x in color]
-        legend_elements = [Patch(facecolor=_COLOR_A, edgecolor=_edge_color(_COLOR_A), label='$a$'),
-                           Patch(facecolor=_COLOR_B, edgecolor=_edge_color(_COLOR_B), label='$b$'),
-                           Patch(facecolor=_COLOR_C, edgecolor=_edge_color(_COLOR_C), label='$c$')]
+        legend_elements = [Patch(facecolor=COLOR_A, edgecolor=_edge_color(COLOR_A), label='$a$'),
+                           Patch(facecolor=COLOR_B, edgecolor=_edge_color(COLOR_B), label='$b$'),
+                           Patch(facecolor=COLOR_C, edgecolor=_edge_color(COLOR_C), label='$c$')]
         if all_mixes or data is not None:
             def add_optional_patch(color, label, insertion=False):
                 if all_mixes or (data is not None
@@ -523,11 +399,11 @@ class TernaryAxesSubplotPoisson(ternary.TernaryAxesSubplot):  # pragma: no cover
                         legend_elements.insert(0, patch)
                     else:
                         legend_elements.append(patch)
-            add_optional_patch(_COLOR_NO_ONE, 'No one', insertion=True)
-            add_optional_patch(_COLOR_AB, 'a, b')
-            add_optional_patch(_COLOR_AC, 'a, c')
-            add_optional_patch(_COLOR_BC, 'b, c')
-            add_optional_patch(_COLOR_ABC, 'a, b, c')
+            add_optional_patch(COLOR_NO_ONE, 'No one', insertion=True)
+            add_optional_patch(COLOR_AB, 'a, b')
+            add_optional_patch(COLOR_AC, 'a, c')
+            add_optional_patch(COLOR_BC, 'b, c')
+            add_optional_patch(COLOR_ABC, 'a, b, c')
         self.legend(title=title, handles=legend_elements)
 
     @staticmethod
@@ -545,7 +421,7 @@ class TernaryAxesSubplotPoisson(ternary.TernaryAxesSubplot):  # pragma: no cover
         of the legend, for example, are not meant to work here.
         """
         legend_ax = plt.gcf().add_axes([.68, .63, .2, .2], facecolor='k')
-        legend_ax.imshow(_DATA_PALETTE, origin='lower', interpolation='gaussian')
+        legend_ax.imshow(DATA_PALETTE, origin='lower', interpolation='gaussian')
         legend_ax.set_xticks([])
         legend_ax.set_yticks([])
         legend_ax.axis('off')